Uploaded V1.0
Has implementations and testing for Bubble Sort, Selection Sort, Insertion Sort, Merge Sort, and Quick sort. As well as prints. XML documentation is fairly in-depth and mostly complete for almost every bit of code, but there's still some work to be done. Current build only allows for code to be ran and whatever was last set in main is what is what is tested (plan to change this).